converted valuePer100UnitServing and portionValue vars from float32 to float64

+++ model/foods.go
@@ -38,8 +38,8 @@ type NutrientReportRequest struct {
 	Nutrient  int     `json:"nutrientno" binding:"required"`
 	FoodGroup string  `json:"foodGroup,omitEmpty"`
 	Sort      string  `json:"sort,omitEmpty"`
-	ValueGTE  float32 `json:"valueGTE"`
-	ValueLTE  float32 `json:"valueLTE"`
+	ValueGTE  float64 `json:"valueGTE"`
+	ValueLTE  float64 `json:"valueLTE"`
 }
 
 // SearchRequest wraps a POST search
@@ -162,9 +162,9 @@ type NutrientData struct {
 	Category     string      `json:"category,omitempty"`
 	Source       string      `json:"Datasource"`
 	Type         string      `json:"type"`
-	Value        float32     `json:"valuePer100UnitServing"`
+	Value        float64     `json:"valuePer100UnitServing"`
 	Portion      string      `json:"portion,omitempty"`
-	PortionValue float32     `json:"portionValue,omitempty"`
+	PortionValue float64     `json:"portionValue,omitempty"`
 	Unit         string      `json:"unit"  binding:"required"`
 	Derivation   *Derivation `json:"derivation,omitempty"`
 	Nutrientno   uint        `json:"nutrientNumber"`
@@ -187,12 +187,12 @@ type NutrientFoodBrowse struct {
 
 // NutrientFoodBrowseItem is the list of nutrient data returned by the food nutrient endpoints
 type NutrientFoodBrowseItem struct {
-	Value        float32     `json:"valuePer100UnitServing"`
+	Value        float64     `json:"valuePer100UnitServing"`
 	Unit         string      `json:"unit"  binding:"required"`
 	Derivation   *Derivation `json:"derivation,omitempty"`
 	Nutrientno   uint        `json:"nutrientNumber"`
 	Nutrient     string      `json:"nutrientName"`
-	PortionValue float32     `json:"valuePerPortion"`
+	PortionValue float64     `json:"valuePerPortion"`
 }
 
 // NutrientReportData is an item returned in a nutrient report
@@ -200,9 +200,9 @@ type NutrientReportData struct {
 	FdcID           string  `json:"fdcId" binding:"required"`
 	Upc             string  `json:"upc"`
 	FoodDescription string  `json:"foodDescription"`
-	Value           float32 `json:"valuePer100UnitServing"`
+	Value           float64 `json:"valuePer100UnitServing"`
 	Portion         string  `json:"portion,omitempty"`
-	PortionValue    float32 `json:"valuePerPortion"`
+	PortionValue    float64 `json:"valuePerPortion"`
 	Unit            string  `json:"unit"`
 	Type            string  `json:"type"`
 }
